update and optimization code

+++ Core/Application/Models/DTOs/Food/AddFoodRequest.cs
@@ -6,7 +6,7 @@ public class AddFoodRequest
 {
     public string Name { get; set; }
     public string Description { get; set; }
-    public uint Price { get; set; }
+    public float Price { get; set; }
     public string RestaurantId { get; set; }
     public List<string> CategoryIds { get; set; }
     public IFormFile File { get; set; }
