+++ Core/Application/Models/DTOs/Food/FoodInfoDto.cs
@@ -5,7 +5,7 @@ public class FoodInfoDto
     public string Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public uint Price { get; set; }
+    public float Price { get; set; }
     public List<string> CategoryIds { get; set; }
     public string ImageUrl { get; set; }
 }
