+++ Core/Application/Models/DTOs/Order/InfoOrderDto.cs
@@ -11,6 +11,6 @@ public class InfoOrderDto
     public bool PayedWithCard { get; set; }
     public string UserId { get; set; }
     public OrderStatus OrderStatus { get; set; }
-    public uint Amount { get; set; }
+    public float Amount { get; set; }
     public byte Rate { get; set; }
 }
