+++ Core/Domain/Models/Order.cs
@@ -14,7 +14,7 @@ public class Order : BaseEntity
     public List<string> OrderedFoodIds { get; set; }
     public DateTime OrderDate { get; set; }
     public DateTime OrderFinishTime { get; set; }
-    public uint Amount { get; set; }
+    public float Amount { get; set; }
     public bool PayedWithCard { get; set; }
     public bool IsActivated { get; set; } 
     public OrderStatus OrderStatus { get; set; }
